Refactor: custom hook 분리
