+++ Week 04/assignment/src/pages/Login/Login.tsx
@@ -1,36 +1,10 @@
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { Wrapper, Title } from './Login.style';
 import { Form, Input } from '../../components/SignupForm/Form.style';
-import { useRef, useState } from 'react';
-import { postLogin } from '../../apis/user';
+import useLogin from '../../hooks/useLogin';
 
 const Login = () => {
-  const navigate = useNavigate();
-  const idRef = useRef<HTMLInputElement | null>(null);
-  const pwdRef = useRef<HTMLInputElement | null>(null);
-  const [errmeg, setErrmeg] = useState<string>('');
-
-  const handleClick = async () => {
-    if (!idRef.current?.value || !pwdRef.current?.value) {
-      setErrmeg('아이디와 비밀번호를 모두 입력해주세요!');
-      return;
-    }
-
-    const prop = { username: idRef.current.value, password: pwdRef.current.value };
-    const res = await postLogin(prop);
-    switch (res) {
-      case '01':
-        setErrmeg('비밀번호가 일치하지 않습니다.');
-        break;
-      case '02':
-        setErrmeg('아이디가 일치하지 않습니다.');
-        break;
-      default:
-        setErrmeg('');
-        localStorage.setItem('token', res);
-        navigate('/mypage');
-    }
-  };
+  const { idRef, pwdRef, errmeg, handleClick } = useLogin();
   return (
     <Wrapper>
       <Title>로그인</Title>
